==> Code/heritability/03_h2fecund.Rmd <==
---
title: "h2fecund"
author: "Enoch Ng'oma"
date: "3/6/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
```

# Read in data

```{r}
load(file= "../../Data/Processed/predicted_egg_counts.rda")
h2life <- read.table('../../Data/Processed/lifespan_correctedData.txt',
                     sep = "\t", header = TRUE,
                     stringsAsFactors = FALSE)
```

# Set negative predicted egg counts to 0

```{r}
M$predicted_count_linear[M$predicted_count_linear < 0] <- 0

train <- M %>% filter(training_set == "yes")
cor(train$handcount, train$predicted_count_linear, use = "complete.obs")
```

# Merge predicted counts and hand counts

```{r}
# hand count preferred where it occurs
M$egg_count <- if_else(!is.na(M$handcount), 
                       M$handcount, M$predicted_count_linear)

# check
#M[1:20, c(1,3,25:27)]
#M[201:220, c(1,3,25:27)]
```

# Replace NstartF for id S11D33_a_LY from 11 to 15 (see ProjectNotes.md)

```{r}
# create "id" column from fID and treat
M <- unite(M, "id", fID, treat, sep = "_", remove=FALSE)

# needed for code to run
M$setDate <- as.character(M$setDate)
M$flipDate <- as.character(M$flipDate)

# change NstartF for all S11D33_a_LY cases to 15
M <- M %>% 
  mutate(NstartF = replace(NstartF, which(id=="S11D33_a_LY"), 15))


#write_csv(M, path = "../../Data/Processed/feclife_with-image-ids.csv")

#checks
#subset(M, M$id=="S11D33_a_LY")
#subset(M, M$id=="S19D55_a_HS")
```

# Merge with lifespan to get number of females

```{r}
# change back to date format
M$setDate <- as.Date(M$setDate)
M$flipDate <- as.Date(M$flipDate)

ureps <- unique(h2life$id)

h2life$n_females <- rep(NA, nrow(h2life))

h2life.t <- h2life[0,]

for(k in ureps) {
  h2life.s <- subset(h2life, id==k)
  for(i in 1:nrow(h2life.s)) {
    if(i == 1) {
      h2life.s[i,'n_females'] <- h2life.s$NstartF[1]
    } else {
      if(is.na(h2life.s$deadF[i])|is.na(h2life.s$carriedF[(i-1)])) {
        h2life.s[i,'n_females'] <- h2life.s[(i-1),'n_females']
      } else {
        ndead <- h2life.s[i,'deadF']-h2life.s[(i-1),'carriedF']
        h2life.s[i,'n_females'] <- h2life.s[(i-1),'n_females']-ndead
      }#else close
    }#else close
  }#for i close
  h2life.t <- rbind(h2life.t, h2life.s)
}#for k close

h2life <- h2life.t

min(h2life$n_females,na.rm=TRUE)
which(is.na(h2life$n_females))
# h2life[which(h2life$n_females<0),]
hist(h2life[which(h2life$n_females<0),'n_females'])


# set negative n_females values to zero
h2life[h2life$n_females<0,'n_females']<-0
h2life[which(h2life$n_females<0),]


# pull all Mondays from 2/29 to 6/20
# females present on each Monday are gave the eggs collected each Tuesday
Mondays <- as.Date("2016-02-29") #YYYY-MM-DD
weekdays(Mondays)
xm <- seq(Mondays, by="7 days", length.out=17)


# pull rows holding Monday flips
dd <- subset(h2life, flipDate %in% as.character(xm))

# Overwrite flipDate with +1 day to match M
dd$flipDate <- as.character(as.Date(dd$flipDate) + 1)

M$NewAge <- as.numeric(M$flipDate - M$setDate)+2 # for checking join

M$setDate <- as.character(M$setDate)
M$flipDate <- as.character(M$flipDate)
```

```{r}
# Join M and dd on flipDate and id.
# Keep only columns that we will use later.
D <- right_join(M, dd, by = c("flipDate", "id")) %>%
  dplyr::select(-contains(".y")) %>% 
  dplyr::select(-box.x, -ccoord.x, -rcoord.x,
                -X__1, -handcount, -handcounted,
                -visually_recheck, -area,
                -predicted_count_linear,
                -training_set)

# Strip out ".x" from column names
names(D) <- str_replace_all(names(D), ".x", "")
```

# Drop rows where egg_count == 0

```{r}
D <- D %>% 
  drop_na(egg_count)
```

# Add column for fecundity per female. Set to 0 if no females.

```{r}
D <- D %>% 
  mutate(eggs_per_female = if_else(n_females == 0,
                                   0, egg_count / n_females),
         cameraid = cameraid)

D %>% 
  ggplot(aes(eggs_per_female)) + geom_histogram()

D %>% 
  filter(eggs_per_female > 100) %>% 
  select(cameraid, id, setDate, flipDate, egg_count, n_females, eggs_per_female) %>% 
  arrange(desc(eggs_per_female)) %>% 
  as.data.frame()

#write_csv(D, path = "../../Data/Processed/eggs_per_female.csv")
```

# IMG_2141.JPG appeares just once for 2016-03-15. Merge seems to spawn/recycle a second event at end.
### resolved

```{r}
#D[D$id=="S17D50_a_STD",]
#D[D$cameraid=="IMG_2141.JPG",]
#D[D$cameraid=="IMG_2141.JPG","flipDate"]

max(D$eggs_per_female)
which.max(D$eggs_per_female)
D[which.max(D$eggs_per_female),]

tt <- subset(D, id=="S17D50_a_STD")
select(tt, 1:5,16:24)

# Image IMG_5532.JPG with 277 eggs/female actually has 2 eggs
D[D$cameraid=="IMG_5532.JPG","eggs_per_female"]<-2

# IMG_3626.JPG, S21D63_a_STD, flipDate: 2016-04-26 with 397 eggs: set to missing (odd case)
D[D$cameraid=="IMG_3626.JPG","eggs_per_female"] <- NA
D[D$cameraid=="IMG_5736.JPG","eggs_per_female"] <- NA
D[D$cameraid=="IMG_5735.JPG","eggs_per_female"] <- NA
D[D$cameraid=="IMG_5738.JPG","eggs_per_female"] <- NA
D[D$cameraid=="IMG_3654.JPG","eggs_per_female"] <- NA

select(tt, 1:4,23:24)
ggplot(D, aes(eggs_per_female)) + geom_histogram()
```

```{r}
# Sort by eggs_per_female
D %>% select(cameraid, treat, flipDate, eggs_per_female) %>%
  arrange(desc(eggs_per_female)) %>% 
  slice(1:50) %>% 
  as.data.frame()

write.table(D, file="../../Data/Processed/eggs_per_female.txt",
            sep = '\t', row.names = FALSE)
```

## Aggregate sum of eggs per vial across all weeks

```{r}
D %>% group_by(id) %>% tally() %>% arrange(n)
D %>% group_by(id) %>% tally() %>% arrange(desc(n))
```

```{r}
(ch.f <- D %>% group_by(id) %>%
   summarise('dfem' = n_distinct(NstartF)) %>%
   filter(dfem > 1) %>% 
   as.data.frame())
```

```{r}
getmode <- function(v) {
   uniqv <- unique(v)
   uniqv[which.max(tabulate(match(v, uniqv)))]
}

D_sum <- D %>% group_by(id) %>% 
  summarize(egg_total = sum(egg_count),
            NstartF = getmode(NstartF),
            egg_total = egg_total / NstartF) %>% 
  select(-NstartF)

D_sum <- D_sum %>% 
  mutate(siredam = str_split(id, "_", simplify = TRUE)[, 1],
         repl = str_split(id, "_", simplify = TRUE)[, 2],
         treat = str_split(id, "_", simplify = TRUE)[, 3],
         sireid = str_split(siredam, "D", simplify = TRUE)[, 1],
         damid = paste0("D", str_split(siredam, "D", simplify = TRUE)[, 2])) %>% 
  select(-siredam)

write.table(D_sum, file="../../Data/Processed/eggs_per_vial.txt",
            sep = '\t', row.names = FALSE)
```
