==> Code/heritability/04_h2fec_visualization.Rmd <==
---
title: "halfsib_fecundity"
author: "Enoch Ng'oma"
date: "3/26/2018"
output:
  html_document:
    theme: flatly
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: true
---

# Load packages

```{r}
library(tidyverse) 
library(cowplot)
library(forcats)
```

# Function to calculate the mean and the standard deviation for each group
  
```{r}
# data : a data frame
# varname : name of column to be summarized
# groupnames : vector of grouping variables

data_summary <- function(data, varname, groupnames){
  summary_func <- function(x, col){
    c(mean = mean(x[[col]], na.rm=TRUE),
      sd = sd(x[[col]], na.rm=TRUE))
  }
  data_sum <- data %>%
    group_by(groupnames) %>% 
    summarize(summary_func(., varname))
  data_sum <- rename(data_sum, c("mean" = varname))
 return(data_sum)
}
```

# Load functions for color mapping

```{r}
source("color_map.R")
source("ggplot_theme.R")
```

# Load data

```{r}
h2fec_phenotype <- read.table('../../Data/Processed/eggs_per_vial.txt',
                              sep = "\t", header = TRUE,
                              stringsAsFactors = FALSE) %>% 
  rename(Diet = treat) %>% 
  mutate(Diet = case_when(
    Diet == "HS" ~ "HS",
    Diet == "LY" ~ "DR",
    Diet == "STD" ~ "C"),
    Diet = factor(Diet),
    Diet = fct_relevel(Diet, "C"))
```

# Density

```{r, dens}
ggplot(h2fec_phenotype, aes(egg_total, fill = Diet)) + 
  geom_density(alpha = 0.5) +
  scale_x_continuous(expand = c(0, 0), limits = c(0, 550)) +
  scale_y_continuous(expand = c(0, 0)) +
  theme(axis.title = element_text(size = 16)) +
  xlab("Total fecundity per female") +
  ylab("Density") +
  theme(legend.position=c(0.8, 0.85)) +
  tx_fill_map() +
  my_theme
ggsave("../../Figures/density_all-fec_data.pdf", width = 6, height = 4)
```

# Summarize data for each diet

```{r}
h2fec_sum <- h2fec_phenotype %>% 
  group_by(Diet) %>% 
  summarise(egg_total_mean = mean(egg_total, na.rm = TRUE),
            sd = sd(egg_total)) %>% 
  rename(egg_total = egg_total_mean)
h2fec_sum
```

# Diet groups and error bars

```{r}
ggplot(data=h2fec_phenotype, aes(x = Diet, y = egg_total, color = Diet)) +
  geom_jitter(position = position_jitter(width = 0.4), alpha = 0.7) +
  tx_color_map() +
  labs(x = "Diet", y = "Total eggs") + 
  expand_limits(y = c(1, 5e2)) 
``` 

# Summarize data for each sire family in each diet

```{r}
# new grouping variable
h2fec_phenotype <- unite(h2fec_phenotype, "sire_Diet", sireid,
                         Diet, sep = "_", remove = FALSE)  

#ascending order
ggplot(h2fec_phenotype, aes(x = reorder(sire_Diet, egg_total),
                            y = egg_total, fill = Diet,
                            color = Diet)) +
  theme(axis.text.x = element_blank(), legend.position = 'top') +
  stat_summary(fun.y = mean,
               position = position_dodge(width = 0.9),
               geom = "bar", alpha = 0.5) +
  geom_point(position = position_dodge(width = 0.9),
             size = 0.75) +
  xlab("Sire family") +
  ylab("Mean fecundity") +
  scale_y_continuous(expand = c(0, 0)) +
  tx_fill_map() +
  tx_color_map() +
  expand_limits(y = c(1, 5e2)) +
  my_theme

ggsave("../../Figures/mean_family_fecundity.pdf", width = 6, height = 4)
```

# Reaction norms for sire families

```{r}
# prep each treatment into its own column

#pull HS and compute median
HS <- subset(h2fec_phenotype, Diet == "HS")
hs_trim <- cbind.data.frame(HS$sireid, HS$egg_total)
hs_med <- aggregate(. ~ HS$sireid, data = hs_trim, FUN = median)
hs_med[, 2] <- NULL
colnames(hs_med) <- c("sireid", "eggs")
hs_med$diet <- "HS"

#pull DR and compute median
DR <- subset(h2fec_phenotype, Diet == "DR")
dr_trim <- cbind.data.frame(DR$sireid, DR$egg_total)
dr_med <- aggregate(. ~ DR$sireid, data = dr_trim, FUN = median)
dr_med[, 2] <- NULL
colnames(dr_med) <- c("sireid", "eggs")
dr_med$diet <- "DR"

#pull C and compute median
C <- subset(h2fec_phenotype, Diet == "C")
c_trim <- cbind.data.frame(C$sireid, C$egg_total)
c_med <- aggregate(. ~ C$sireid, data = c_trim, FUN = median)
c_med[, 2] <- NULL
colnames(c_med) <- c("sireid", "eggs")
c_med$diet <- "C"

# Join data frames
B <- rbind(hs_med, c_med, dr_med) %>% 
  mutate(diet = factor(diet, levels = c("HS", "C", "DR")))
```

# plot reaction norms for sire families

```{r}
ggplot(B, aes(x = diet, y = eggs, color = sireid)) +
  geom_point(size = 3) +
  geom_line(aes(group = sireid), alpha = 0.5, size = 1) +
  labs(x = "Diet", y = "Total eggs") +
  theme(legend.position = "none") +
  coord_cartesian(xlim = c(1.5, 2.5)) +
  my_theme

Fecundity_GxE <- last_plot()
ggsave("../../Figures/h2fec_reaction_norms.pdf", width = 6, height = 4)
save(Fecundity_GxE, file = "../../Figures/Fecundity_GxE.Rda")
```

# Per female analysis

```{r}
AA <- read.table('../../Data/Processed/eggs_per_female.txt',
                 sep = "\t", header = TRUE,
                 stringsAsFactors = FALSE) %>% 
  select(flipDate, id, fID, treat, age, sireid, damid,
         n_females, eggs_per_female)

# Prepare data
AA <- AA %>%
  rename(Diet = treat) %>%
  mutate(Diet=replace(Diet, Diet=="LY", "DR")) %>%
  mutate(Diet=replace(Diet, Diet=="STD", "C"))

AA$flipDate <- as.Date(AA$flipDate)

AA <- AA[!is.na(AA$eggs_per_female), ]

# Add sireXtreatment grouping variable
AA <- unite(AA, "sire_trt", sireid, Diet, sep = "_", remove=FALSE)
A2 <- aggregate(AA["eggs_per_female"], by = AA["sire_trt"], sum)
```

# Plot mean age per female per day

```{r, fig.height=4, fig.width=12}
ggplot(AA, aes(x = age, y = eggs_per_female, color = Diet)) +
  geom_point(position = position_jitter(width = 0.3), alpha = 0.7) +
  tx_color_map() +
  xlab("Age (days)") +
  ylab("Mean eggs/female/day") +
  geom_smooth(method = "lm", se = TRUE, color="black")+
  theme(legend.position=c(0.95, 0.85)) +
  facet_grid(. ~ Diet) +
  my_theme
ggsave("../../Figures/h2_avg_fecundity_per_female_per_day.pdf",
       width = 12, height = 4)
```

```{r}
AA %>% 
  ggplot(aes(x = age, y = eggs_per_female, color = Diet, group = fID)) +
  geom_line(alpha = 0.25) +
  geom_smooth(aes(group = Diet), se = FALSE, color = "black") +
  tx_color_map() +
  facet_grid(. ~ Diet) +
  labs(x = "Age (days)", y = "Eggs per Female") +
  theme(legend.position = "none") +
  my_theme

Fec_line_plot <- last_plot()
save(Fec_line_plot, file = "../../Figures/Fec_line_plot.Rda")
```

```{r}
AA%>%
group_by(Diet)%>% 
summarise_each(funs(max, min, mean, median, sd), eggs_per_female)
```
