==> Code/heritability/05_h2surv_analysis.Rmd <==
---
title: "Analysis of heritability of a fly half-sibling experiment"
author: "Kevin Middleton"
date: "3/7/2017"
output:
  html_document:
    theme: flatly
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: true
---

```{r}
library(MCMCglmm)
library(rethinking)
library(tidyverse)
library(cowplot)

source("color_map.R")
source("ggplot_theme.R")
```

# Heritability

```{r eval=TRUE}
set.seed(37264)

iter <- 2e6
burnin <- 15000
rerun_MCMC <- FALSE

h2life <- read.table('../../Data/Processed/Female_events_lifespan.txt',
                     sep = "\t", header = TRUE,
                     stringsAsFactors = FALSE)

h2_filtered <- h2life[h2life$status!=3, ] # remove censored events
h2_filtered$animal <- seq(1, nrow(h2_filtered))
h2_filtered$treat <- as.factor(h2_filtered$treat)
```

## High sugar, HS

```{r}
HS <- subset(h2_filtered, treat == "HS")

pedigree <- HS[, c("animal", "sireid", "damid")]
names(pedigree) <- c("animal", "sire", "dam")
pedigree$animal <- as.character(pedigree$animal)
sires <- data.frame(animal = unique(pedigree$sire),
                    sire = NA, dam = NA, stringsAsFactors = FALSE)
dams <- data.frame(animal = unique(pedigree$dam),
                   sire = NA, dam = NA, stringsAsFactors = FALSE)
pedigree <- bind_rows(sires, dams, pedigree) %>% as.data.frame()
```

```{r}
prior <- list(R = list(V = 1, nu = 0.002),
              G = list(G1 = list(V = 1, nu = 0.002)))

if (rerun_MCMC) {
  model <- MCMCglmm(NewAge ~ 1,
                    random = ~ animal,
                    family = "gaussian",
                    prior = prior,
                    pedigree = pedigree,
                    data = HS,
                    nitt = iter,
                    burnin = burnin,
                    thin = 50,
                    verbose = TRUE)
  
  save(model, file = "../../Data/Processed/HS_lifespan.Rda")
}
```

```{r}
load("../../Data/Processed/HS_lifespan.Rda")

# Fixed
plot(model$Sol)
autocorr.diag(model$Sol)
effectiveSize(model$Sol)

# Random
plot(model$VCV)
autocorr.diag(model$VCV)
effectiveSize(model$VCV)

summary(model)

herit <- model$VCV[, "animal"] / (model$VCV[, "animal"] + model$VCV[, "units"])
plot(herit)
median(herit)
f <- HPDinterval(herit)
k <- as.data.frame(f)
```

```{r}
density(herit)

hs.samp <- data.frame(('h2' = herit))

ggplot(hs.samp, aes(x = var1), y = k) +
  geom_histogram() +
  geom_vline(xintercept = median(herit))

rm(model) 
```

## Low yeast, DR

```{r}
DR <- subset(h2_filtered, treat == "LY")

pedigree <- DR[, c("animal", "sireid", "damid")]
names(pedigree) <- c("animal", "sire", "dam")
pedigree$animal <- as.character(pedigree$animal)
sires <- data.frame(animal = unique(pedigree$sire),
                    sire = NA, dam = NA, stringsAsFactors = FALSE)
dams <- data.frame(animal = unique(pedigree$dam),
                   sire = NA, dam = NA, stringsAsFactors = FALSE)
pedigree <- bind_rows(sires, dams, pedigree) %>% as.data.frame()

prior <- list(R = list(V = 1, nu = 0.002),
              G = list(G1 = list(V = 1, nu = 0.002)))

if (rerun_MCMC) {
  model <- MCMCglmm(NewAge ~ 1,
                    random = ~ animal,
                    family = "gaussian",
                    prior = prior,
                    pedigree = pedigree,
                    data = DR,
                    nitt = iter,
                    burnin = burnin,
                    thin = 50,
                    verbose = TRUE)
  
  save(model, file = "../../Data/Processed/DR_lifespan.Rda")
}
```

```{r}
load("../../Data/Processed/DR_lifespan.Rda")

# Fixed
plot(model$Sol)
autocorr.diag(model$Sol)
effectiveSize(model$Sol)

# Random
plot(model$VCV)
autocorr.diag(model$VCV)
effectiveSize(model$VCV)

summary(model)

herit.dr <- model$VCV[, "animal"] / (model$VCV[, "animal"] + model$VCV[, "units"])
plot(herit.dr)
median(herit.dr)
HPDinterval(herit.dr)
```

```{r}
density(herit.dr)

dr.samp<-data.frame(('h2' = herit.dr))

ggplot(dr.samp, aes(x = var1), y = k) +
  geom_histogram() +
  geom_vline(xintercept = median(herit.dr))

rm(model)
```

## Standard, STD

```{r}
STD <- subset(h2_filtered, treat == "STD")

pedigree <- STD[, c("animal", "sireid", "damid")]
names(pedigree) <- c("animal", "sire", "dam")
pedigree$animal <- as.character(pedigree$animal)
sires <- data.frame(animal = unique(pedigree$sire),
                    sire = NA, dam = NA, stringsAsFactors = FALSE)
dams <- data.frame(animal = unique(pedigree$dam),
                   sire = NA, dam = NA, stringsAsFactors = FALSE)
pedigree <- bind_rows(sires, dams, pedigree) %>% as.data.frame()

prior <- list(R = list(V = 1, nu = 0.002),
              G = list(G1 = list(V = 1, nu = 0.002)))

if (rerun_MCMC) {
  model <- MCMCglmm(NewAge ~ 1,
                    random = ~ animal,
                    family = "gaussian",
                    prior = prior,
                    pedigree = pedigree,
                    data = STD,
                    nitt = iter,
                    burnin = burnin,
                    thin = 50,
                    verbose = TRUE)
  
  save(model, file = "../../Data/Processed/STD_lifespan.Rda")
}
```

```{r}
load("../../Data/Processed/STD_lifespan.Rda")

# Fixed
plot(model$Sol)
autocorr.diag(model$Sol)
effectiveSize(model$Sol)

# Random
plot(model$VCV)
autocorr.diag(model$VCV)
effectiveSize(model$VCV)

summary(model)

herit.c <- model$VCV[, "animal"] / (model$VCV[, "animal"] +
                                      model$VCV[, "units"])
plot(herit.c)
median(herit.c)
p <- HPDinterval(herit.c)
q <- as.data.frame(f)
```

# Plot STD with ggplot

```{r}
density(herit.c)

c.samp<-data.frame(('h2' = herit.c))

ggplot(c.samp, aes(x = var1), y = k) +
  geom_histogram()+
  geom_vline(xintercept = median(herit.c))
```

# The three files together

```{r, }
hs.samp$Diet <- "HS"
dr.samp$Diet <- "DR"
c.samp$Diet <- "C"

hs.samp$HS <- NULL
dr.samp$DR <- NULL
c.samp$C <- NULL
heritab <- rbind(hs.samp, dr.samp, c.samp)
save(heritab, file = "../../Data/Processed/herit_lifespan.Rda")
```

# Pairwise differences

```{r}
C_DR <- c.samp$var1 - dr.samp$var1
C_HS <- c.samp$var1 - hs.samp$var1
HS_DR <- hs.samp$var1 - dr.samp$var1

HPDinterval(as.mcmc(C_DR))
HPDinterval(as.mcmc(C_HS))
HPDinterval(as.mcmc(HS_DR))
```

# Plot treatments together

```{r, dens}
load("../../Data/Processed/herit_lifespan.Rda")

ggplot(heritab, aes(var1, fill = Diet)) + 
  geom_density(alpha = 0.2) +
  xlab("Heritability") +
  ylab("Density") +
  theme(legend.position = c(0.85, 0.85)) +
  tx_fill_map() +
  my_theme
Lifespan_h2 <- last_plot()
save(Lifespan_h2, file = "../../Figures/Lifespan_h2.Rda")
```
