==> Code/heritability/06_h2fec_heritability.Rmd <==
---
title: "Dmel half-sib heritability in fecundity"
author: "Kevin Middleton"
date: "3/7/2017"
output:
  html_document:
    theme: flatly
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: true
---

```{r}
library(knitr)
# purl("06_h2fec_heritability.Rmd", output = "06_h2fec_early_heritability.R", documentation = 2)
```

```{r}
library(MCMCglmm)
library(tidyverse)
library(rstanarm)
library(cowplot)

source("color_map.R")
source("ggplot_theme.R")
```

# Heritability

```{r}
rerun_MCMC <- FALSE
total_fecundity <- TRUE # early fecundity if FALSE

set.seed(33416)

iter <- 2e6
burnin <- 1.5e4
thin <- 50
```

## Data for early life fecundity

```{r}
AA <- read.table('../../Data/Processed/eggs_per_female.txt',
                 sep = "\t", header = TRUE,
                 stringsAsFactors = FALSE) %>% 
  select(flipDate, id, fID, treat, age, sireid, damid,
         n_females, eggs_per_female) %>%
  mutate(treat = case_when(
    treat == "HS" ~ "HS",
    treat == "LY" ~ "DR",
    treat == "STD" ~ "C"),
    treat = factor(treat),
    treat = fct_relevel(treat, "C")) %>%
  rename(Diet = treat)

# Prepare data

AA$flipDate <- as.Date(AA$flipDate)

AA <- AA[!is.na(AA$eggs_per_female), ]

# Get early life fec - nearest day 5
Day <- 5
all.ids <- unique(AA$id)

early.fec <- data.frame(
  "id" = character(length = length(all.ids)),
  "egg_total" = numeric(length = length(all.ids)),
  "age" = numeric(length = length(all.ids)),
  stringsAsFactors = FALSE)

for(ii in 1:length(all.ids)) {
  sub.dat <- subset(AA, id == all.ids[ii])
  early.fec[ii, 'id'] <- all.ids[ii]
  early.fec[ii, 'egg_total'] <- sub.dat[which.min(abs(sub.dat$age - Day)),
                                       'eggs_per_female']
  early.fec[ii, 'age'] <- sub.dat[which.min(abs(sub.dat$age - Day)),
                                 'age']
}

early.fec <- early.fec %>%
  mutate(siredam = str_split(id, "_", simplify = TRUE)[, 1],
         repl = str_split(id, "_", simplify = TRUE)[, 2],
         treat = str_split(id, "_", simplify = TRUE)[, 3],
         sireid = str_split(siredam, "D", simplify = TRUE)[, 1],
         damid = paste0("D", str_split(siredam, "D", simplify = TRUE)[, 2])) %>% 
  dplyr::select(-siredam)
```

## Select data: total fecundity or early fecundity 

```{r}
if (total_fecundity) {
  h2fec <- read.table("../../Data/Processed/eggs_per_vial.txt",
                    sep = '\t',
                    dec = ".", header = TRUE,
                    stringsAsFactors = FALSE)
  label <- "total_fec"
} else {
  h2fec <- early.fec
  label <- "early_fec"
}
```

## Final data processing

```{r}
# Standardize egg_total
h2fec$egg_total <- as.numeric(scale(h2fec$egg_total))

h2fec$animal <- seq(1, nrow(h2fec))
h2fec$treat <- as.factor(h2fec$treat)
```

# High sugar, HS

```{r}
HS <- subset(h2fec, treat == "HS")

pedigree <- HS[, c("animal", "sireid", "damid")]
names(pedigree) <- c("animal", "sire", "dam")
pedigree$animal <- as.character(pedigree$animal)
sires <- data.frame(animal = unique(pedigree$sire),
                    sire = NA, dam = NA, stringsAsFactors = FALSE)
dams <- data.frame(animal = unique(pedigree$dam),
                   sire = NA, dam = NA, stringsAsFactors = FALSE)
pedigree <- bind_rows(sires, dams, pedigree) %>% as.data.frame()

outfile <- paste0("../../Data/Processed/HS_", label, ".Rda")

if (rerun_MCMC) {
  prior <- list(R = list(V = 1, nu = 0.002),
                G = list(G1 = list(V = 1,
                                   nu = 1,
                                   alpha.mu = 0,
                                   alpha.V = 10)))
  
  model <- MCMCglmm(egg_total ~ 1,
                    random = ~ animal,
                    family = "gaussian",
                    prior = prior,
                    pedigree = pedigree,
                    data = HS,
                    nitt = iter,
                    burnin = burnin,
                    thin = thin,
                    verbose = TRUE)
  save(model, file = outfile)
}
```

```{r HS_analysis}
load(outfile)

# Fixed
plot(model$Sol)
autocorr.diag(model$Sol)
effectiveSize(model$Sol)

# Random
plot(model$VCV)
autocorr.diag(model$VCV)
effectiveSize(model$VCV)

summary(model)

herit.hs <- model$VCV[, "animal"] /
  (model$VCV[, "animal"] + model$VCV[, "units"])
plot(herit.hs)
median(herit.hs)
h <- HPDinterval(herit.hs)
h1 <- as.data.frame(h)
herit.hs <- as.data.frame(herit.hs)
```

# Low yeast, DR

```{r}
outfile <- paste0("../../Data/Processed/DR_", label, ".Rda")

if (rerun_MCMC) {
  DR <- subset(h2fec, treat == "LY")
  
  pedigree <- DR[, c("animal", "sireid", "damid")]
  names(pedigree) <- c("animal", "sire", "dam")
  pedigree$animal <- as.character(pedigree$animal)
  sires <- data.frame(animal = unique(pedigree$sire),
                      sire = NA, dam = NA, stringsAsFactors = FALSE)
  dams <- data.frame(animal = unique(pedigree$dam),
                     sire = NA, dam = NA, stringsAsFactors = FALSE)
  pedigree <- bind_rows(sires, dams, pedigree) %>% as.data.frame()

  prior <- list(R = list(V = 1, nu = 0.002),
                G = list(G1 = list(V = 1,
                                   nu = 1,
                                   alpha.mu = 0,
                                   alpha.V = 10)))
  
  model <- MCMCglmm(egg_total ~ 1,
                    random = ~ animal,
                    family = "gaussian",
                    prior = prior,
                    pedigree = pedigree,
                    data = DR,
                    nitt = iter,
                    burnin = burnin,
                    thin = thin,
                    verbose = TRUE)
  
  save(model, file = outfile)
}
```

```{r DR_analysis}
load(outfile)

# Fixed
plot(model$Sol)
autocorr.diag(model$Sol)
effectiveSize(model$Sol)

# Random
plot(model$VCV)
autocorr.diag(model$VCV)
effectiveSize(model$VCV)

summary(model)

herit.dr <- model$VCV[, "animal"] /
  (model$VCV[, "animal"] + model$VCV[, "units"])
plot(herit.dr)
median(herit.dr)
d <- HPDinterval(herit.dr)
d1 <- as.data.frame(d)
herit.dr <- as.data.frame(herit.dr)
```

##### Standard, STD

```{r}
outfile <- paste0("../../Data/Processed/STD_", label, ".Rda")

if (rerun_MCMC) {
  STD <- subset(h2fec, treat == "STD")
  
  pedigree <- STD[, c("animal", "sireid", "damid")]
  names(pedigree) <- c("animal", "sire", "dam")
  pedigree$animal <- as.character(pedigree$animal)
  sires <- data.frame(animal = unique(pedigree$sire),
                      sire = NA, dam = NA, stringsAsFactors = FALSE)
  dams <- data.frame(animal = unique(pedigree$dam),
                     sire = NA, dam = NA, stringsAsFactors = FALSE)
  pedigree <- bind_rows(sires, dams, pedigree) %>% as.data.frame()

  prior <- list(R = list(V = 1, nu = 0.002),
                G = list(G1 = list(V = 1,
                                   nu = 1,
                                   alpha.mu = 0,
                                   alpha.V = 10)))
  
  model <- MCMCglmm(egg_total ~ 1,
                    random = ~ animal,
                    family = "gaussian",
                    prior = prior,
                    pedigree = pedigree,
                    data = STD,
                    nitt = iter,
                    burnin = burnin,
                    thin = thin,
                    verbose = TRUE)
  
  save(model, file = outfile)
}
```

```{r STD_analysis}
load(outfile)

# Fixed
plot(model$Sol)
autocorr.diag(model$Sol)
effectiveSize(model$Sol)

# Random
plot(model$VCV)
autocorr.diag(model$VCV)
effectiveSize(model$VCV)

summary(model)

herit.std <- model$VCV[, "animal"] /
  (model$VCV[, "animal"] + model$VCV[, "units"])
plot(herit.std)
median(herit.std)
c <- HPDinterval(herit.std)
c1 <- as.data.frame(c)
herit.std <- as.data.frame(herit.std)
```

# The three files together

```{r}
herit.hs$Diet <- "HS"
herit.dr$Diet <- "DR"
herit.std$Diet <- "C"

heritab <- rbind(herit.hs, herit.dr, herit.std)
outfile <- paste0("../../Data/Processed/herit_", label, ".Rda")
save(heritab, file = outfile)
```

# Plot treatments together

```{r}
load(outfile)

Fecundity_h2 <- ggplot(heritab, aes(var1, fill = Diet)) + 
  geom_density(alpha = 0.2) +
  expand_limits(x = c(0, 1.0)) +
  xlab("Heritability") +
  ylab("Density") +
  theme(legend.position = c(0.85, 0.85)) +
  my_theme +
  tx_fill_map()
Fecundity_h2
ggsave(Fecundity_h2,
       file = paste0("../../Figures/Fecundity_heritability_", label, "_plots.pdf"),
       width = 4, height = 4)
save(Fecundity_h2, file = paste0("../../Figures/Fecundity_", label, "_h2.Rda"))
```

# Early fecundity comparison mixed model

```{r, results="hide"}
iter <- 2e4
early.fec <- early.fec %>% 
  mutate(egg_total_c = egg_total - mean(egg_total, na.rm = TRUE))

if (rerun_MCMC) {
  fm1 <- stan_glm(egg_total_c ~ 1, data = early.fec,
                  iter = iter,
                  prior_aux = cauchy(0, 1),
                  prior_intercept = normal(0, 10))
  fm2 <- stan_glm(egg_total_c ~ treat, data = early.fec,
                  iter = iter,
                  prior_aux = cauchy(0, 1),
                  prior_intercept = normal(0, 10),
                  prior = normal(0, 10))
  fm3 <- stan_lmer(egg_total_c ~ treat + (1 | sireid), data = early.fec,
                   iter = iter,
                   prior_aux = cauchy(0, 1),
                   prior_intercept = normal(0, 10),
                   prior = normal(0, 10),
                   prior_covariance = decov(1, 1, 1, 1))
  fm4 <- stan_lmer(egg_total_c ~ treat + (1 | sireid:damid),
                   data = early.fec,
                   iter = iter,
                   prior_aux = cauchy(0, 1),
                   prior_intercept = normal(0, 10),
                   prior = normal(0, 10),
                   prior_covariance = decov(1, 1, 1, 1))
  fm5 <- stan_lmer(egg_total_c ~ treat + (1 | sireid) + (1 | sireid:damid),
                   data = early.fec,
                   iter = iter,
                   prior_aux = cauchy(0, 1),
                   prior_intercept = normal(0, 10),
                   prior = normal(0, 10),
                   prior_covariance = decov(1, 1, 1, 1))
  fm6 <- stan_lmer(egg_total_c ~ treat + (1 | sireid) + (1 | sireid:damid) +
                     (1 | treat:sireid),
                   data = early.fec,
                   control = list(adapt_delta = 0.99),
                   iter = iter,
                   prior_aux = cauchy(0, 1),
                   prior_intercept = normal(0, 10),
                   prior = normal(0, 10),
                   prior_covariance = decov(1, 1, 1, 1))
  
  loo_fm1 <- loo(fm1, k_threshold = 0.7)
  loo_fm2 <- loo(fm2, k_threshold = 0.7)
  loo_fm3 <- loo(fm3, k_threshold = 0.7)
  loo_fm4 <- loo(fm4, k_threshold = 0.7)
  loo_fm5 <- loo(fm5, k_threshold = 0.7)
  loo_fm6 <- loo(fm6, k_threshold = 0.7)
  save(list = c("fm1", "fm2", "fm3", "fm4", "fm5", "fm6",
                "loo_fm1", "loo_fm2", "loo_fm3", "loo_fm4", "loo_fm5",
                "loo_fm6"),
       file = "../../Data/Processed/Fecundity_early_models_bayes.Rda")
} else {
  load("../../Data/Processed/Fecundity_early_models_bayes.Rda")
}
```

```{r}
summary(fm1)
summary(fm2)
summary(fm3)
summary(fm4)
summary(fm5)
summary(fm6)
```

```{r}
compare_models(loos = list(loo_fm1, loo_fm2, loo_fm3, loo_fm4,
                           loo_fm5, loo_fm6))
loo_model_weights(list(loo_fm1, loo_fm2, loo_fm3, loo_fm4,
                       loo_fm5, loo_fm6))
```

# Total fecundity comparison mixed model

```{r, results="hide"}
h2fec <- read.table("../../Data/Processed/eggs_per_vial.txt",
                    sep = '\t',
                    dec = ".", header = TRUE,
                    stringsAsFactors = FALSE)

iter <- 1e4
h2fec <- h2fec %>% 
  mutate(egg_total_c = egg_total - mean(egg_total, na.rm = TRUE))

if (rerun_MCMC) {
  fm1 <- stan_glm(egg_total_c ~ 1, data = h2fec,
                  iter = iter,
                  prior_aux = cauchy(0, 1),
                  prior_intercept = normal(0, 10))
  fm2 <- stan_glm(egg_total_c ~ treat, data = h2fec,
                  iter = iter,
                  prior_aux = cauchy(0, 1),
                  prior_intercept = normal(0, 10),
                  prior = normal(0, 10))
  fm3 <- stan_lmer(egg_total_c ~ treat + (1 | sireid), data = h2fec,
                   iter = iter,
                   prior_aux = cauchy(0, 1),
                   prior_intercept = normal(0, 10),
                   prior = normal(0, 10),
                   prior_covariance = decov(1, 1, 1, 1))
  fm4 <- stan_lmer(egg_total_c ~ treat + (1 | sireid:damid),
                   data = h2fec,
                   iter = iter,
                   prior_aux = cauchy(0, 1),
                   prior_intercept = normal(0, 10),
                   prior = normal(0, 10),
                   prior_covariance = decov(1, 1, 1, 1))
  fm5 <- stan_lmer(egg_total_c ~ treat + (1 | sireid) + (1 | sireid:damid),
                   data = h2fec,
                   iter = iter,
                   prior_aux = cauchy(0, 1),
                   prior_intercept = normal(0, 10),
                   prior = normal(0, 10),
                   prior_covariance = decov(1, 1, 1, 1))
  fm6 <- stan_lmer(egg_total_c ~ treat + (1 | sireid) + (1 | sireid:damid) +
                     (1 | treat:sireid),
                   data = h2fec,
                   control = list(adapt_delta = 0.99),
                   iter = iter,
                   prior_aux = cauchy(0, 1),
                   prior_intercept = normal(0, 10),
                   prior = normal(0, 10),
                   prior_covariance = decov(1, 1, 1, 1))
  
  loo_fm1 <- loo(fm1, k_threshold = 0.7)
  loo_fm2 <- loo(fm2, k_threshold = 0.7)
  loo_fm3 <- loo(fm3, k_threshold = 0.7)
  loo_fm4 <- loo(fm4, k_threshold = 0.7)
  loo_fm5 <- loo(fm5, k_threshold = 0.7)
  loo_fm6 <- loo(fm6, k_threshold = 0.7)
  save(list = c("fm1", "fm2", "fm3", "fm4", "fm5", "fm6",
                "loo_fm1", "loo_fm2", "loo_fm3", "loo_fm4", "loo_fm5",
                "loo_fm6"),
       file = "../../Data/Processed/Fecundity_total_models_bayes.Rda")
} else {
  load("../../Data/Processed/Fecundity_total_models_bayes.Rda")
}
```

```{r}
summary(fm1)
summary(fm2)
summary(fm3)
summary(fm4)
summary(fm5)
summary(fm6)
```

```{r}
compare_models(loos = list(loo_fm1, loo_fm2, loo_fm3, loo_fm4,
                           loo_fm5, loo_fm6))
loo_model_weights(list(loo_fm1, loo_fm2, loo_fm3, loo_fm4,
                       loo_fm5, loo_fm6))
```
