==> Code/heritability/09_Two_trait.Rmd <==
---
title: "Two_trait"
author: "EGK"
date: "5/25/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(cowplot)
library(forcats)
library(survival)

source("color_map.R")
source("ggplot_theme.R")

```

## Load Data

```{r}
h2fec_phenotype <- read.table("../../Data/Processed/eggs_per_vial.txt",
                              sep = "\t", header = TRUE,
                              stringsAsFactors = FALSE) %>% 
  rename(Diet = treat) %>% 
  mutate(Diet = case_when(
    Diet == "HS" ~ "HS",
    Diet == "LY" ~ "DR",
    Diet == "STD" ~ "C"),
    Diet = factor(Diet),
    Diet = fct_relevel(Diet, "C"))

h2fec_phenotype$sd_id <- paste(h2fec_phenotype$sireid, h2fec_phenotype$damid, sep="_")

tt <- as.data.frame(table(h2fec_phenotype$sd_id, h2fec_phenotype$Diet))

cc <- unique(h2fec_phenotype[,c('sireid','damid')],MARGIN=1)
cc <- as.data.frame(table(cc$sireid))
cc <- merge(cc, h2fec_phenotype[,c('sireid','damid')], by.x='Var1',by.y='sireid')
cc$sd_id <- paste(cc$Var1,cc$damid, sep="_")

all.count<-merge(tt, cc, by.x='Var1', by.y='sd_id')
colnames(all.count)<-c('sd_id', 'Diet', 'Nreplicates', 'Sire','NDamsPerSire','Dam') 
all.count <- all.count[,c(4,6,2,3,5)]
all.count <- unique(all.count, MARGIN=1)
write.table(all.count, "../../Data/Processed/Sample_counts.txt",sep="\t", row.names=FALSE, quote=FALSE)


tt <- table(h2fec_phenotype$sireid, h2fec_phenotype$Diet)
length(which(rowSums(tt)>17))


h2life <- read.table('../../Data/Processed/Female_events_lifespan.txt',
                     sep = "\t", header = TRUE,
                     stringsAsFactors = FALSE) %>% 
  mutate(treat = case_when(
    treat == "HS" ~ "HS",
    treat == "LY" ~ "DR",
    treat == "STD" ~ "C"),
    treat = factor(treat),
    treat = fct_relevel(treat, "C"))

# Per female analysis
AA <- read.table('../../Data/Processed/eggs_per_female.txt',
                 sep = "\t", header = TRUE,
                 stringsAsFactors = FALSE) %>% 
  select(flipDate, id, fID, treat, age, sireid, damid,
         n_females, eggs_per_female) %>%
  mutate(treat = case_when(
    treat == "HS" ~ "HS",
    treat == "LY" ~ "DR",
    treat == "STD" ~ "C"),
    treat = factor(treat),
    treat = fct_relevel(treat, "C")) %>%
  rename(Diet = treat)


# Prepare data

AA$flipDate <- as.Date(AA$flipDate)

AA <- AA[!is.na(AA$eggs_per_female), ]

# Add sireXtreatment grouping variable
AA <- unite(AA, "sire_trt", sireid, Diet, sep = "_", remove=FALSE)
A2 <- aggregate(AA["eggs_per_female"], by = AA["sire_trt"], sum)
```


## Summarize 

```{r}
km_sire <- survfit(Surv(NewAge, status == 2) ~ id,
                   conf.type = "log", 
                   conf.int = 0.95,
                   type = "kaplan-meier",
                   error = "greenwood",
                   data = h2life)
med.dat<-summary(km_sire)$table[,'median']

ids<-strsplit(names(med.dat), split="=", fixed=TRUE)
ids<-unlist(lapply(ids, function(x) x[2]))
ids_t <- ids
ids<-strsplit(ids, split="_",fixed=TRUE)

med.life <- data.frame(
  'id' = ids_t,
  'Family' = unlist(lapply(ids, function(x) x[1])),
  'Replicate' = unlist(lapply(ids, function(x) x[2])),
  'Treatment' = unlist(lapply(ids, function(x) x[3])),
  'Median_Lifespan' = med.dat,
  stringsAsFactors = FALSE)

# Get early life fec
Day <- 5
all.ids <- unique(AA$id)

early.fec <- data.frame("id" = character(length=length(all.ids)),
                        "earlyegg"=numeric(length=length(all.ids)),
                        "age"=numeric(length=length(all.ids)),                                           stringsAsFactors = FALSE)

for(ii in 1:length(all.ids))
{
  sub.dat <- subset(AA, id==all.ids[ii])
  early.fec[ii,'id']<-all.ids[ii]
  early.fec[ii,'earlyegg']<-sub.dat[which.min(abs(sub.dat$age-Day)),'eggs_per_female']
  early.fec[ii,'age']<-sub.dat[which.min(abs(sub.dat$age-Day)),'age']
  
}



all.dat <- left_join(med.life, h2fec_phenotype)

early.dat <- left_join(all.dat, early.fec)

#get means by diet
sum.dat <- early.dat %>% 
  group_by(Diet) %>% 
  summarise(life = mean(Median_Lifespan),
            fec = mean(earlyegg),
            lifese = sd(Median_Lifespan) / sqrt(length(Median_Lifespan)),
            fecse =sd(earlyegg)/sqrt(length(earlyegg)))


```

```{r}
all.dat %>% 
  ggplot(aes(x = Median_Lifespan, y = egg_total, color = Diet)) +
  geom_point() +
  geom_smooth(se = FALSE, method = "lm") +
  tx_color_map()

ttrait <-ggplot() +
  geom_point(data = early.dat, aes(x = Median_Lifespan, y = earlyegg, color = Diet)) +
  geom_point(data=sum.dat, aes(life, fec, fill=Diet), size=4, pch=21, color='black') +
  geom_errorbar(data=sum.dat, aes(x=life, ymin=fec-fecse, ymax=fec+fecse), width=1) +
  geom_errorbarh(data=sum.dat, aes(x=life, y=fec, xmin=life-lifese, xmax=life+lifese), height=1) +
  #geom_smooth(se = FALSE, method = "lm") +
  ylab("Early life fecundity per female") +
  xlab("Median lifespan (days)") +
  tx_color_map() +
  tx_fill_map() +
  theme(legend.position="none")+
  my_theme

save(ttrait, file="../../Figures/ttrait_plot.rda")

```
